+++ reflex_django_tests/django_settings.py
@@ -24,6 +24,7 @@ MIDDLEWARE = [
     "django.middleware.csrf.CsrfViewMiddleware",
     "django.contrib.auth.middleware.AuthenticationMiddleware",
     "django.contrib.messages.middleware.MessageMiddleware",
+    "reflex_django.streaming_middleware.AsyncStreamingMiddleware",
 ]
 
 ROOT_URLCONF = "reflex_django.urls"