+++ src/reflex_django/default_settings.py
@@ -204,6 +204,7 @@ MIDDLEWARE = [
     "django.middleware.csrf.CsrfViewMiddleware",
     "django.contrib.auth.middleware.AuthenticationMiddleware",
     "django.contrib.messages.middleware.MessageMiddleware",
+    "reflex_django.streaming_middleware.AsyncStreamingMiddleware",
 ]
 
 ROOT_URLCONF = os.environ.get("REFLEX_DJANGO_URLCONF", "reflex_django.urls")